+++ packages/types/src/factors.ts
@@ -1,3 +1,5 @@
+import type { PhoneCodeChannel } from 'phoneCodeChannel';
+
 import type { PublicKeyCredentialWithAuthenticatorAssertionResponse } from './passkey';
 import type {
   BackupCodeStrategy,
@@ -35,6 +37,7 @@ export type PhoneCodeFactor = {
   safeIdentifier: string;
   primary?: boolean;
   default?: boolean;
+  channel?: PhoneCodeChannel;
 };
 
 export type Web3SignatureFactor = {