+++ packages/react-router/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.4.3
+
+### Patch Changes
+
+- Updated dependencies [[`1ff6d6e`](https://github.com/clerk/javascript/commit/1ff6d6efbe838b3f7f6977b2b5215c2cafd715f6), [`fbf3cf4`](https://github.com/clerk/javascript/commit/fbf3cf4916469c4e118870bf12efca2d0f77d9d8)]:
+  - @clerk/shared@3.9.0
+  - @clerk/types@4.58.1
+  - @clerk/clerk-react@5.31.3
+  - @clerk/backend@1.32.1
+
 ## 1.4.2
 
 ### Patch Changes